+++ cli/src/commands/create/quota.rs
@@ -63,9 +63,6 @@ pub fn create(client: &Client, args: &CreateQuotaArgs) -> Result<()> {
         )
         .context("Operation to set quota has failed")?;
 
-    info!(
-        "New quota `{}` set successfully in tenant with id `{}`",
-        tenant_quota_kind, tenant_id
-    );
+    info!("New quota `{tenant_quota_kind}` set successfully in tenant with id `{tenant_id}`");
     Ok(())
 }