+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased 
 - Add ability to get quotas for tenant other than current one 
+- Refresh user permissions every run 
+- Wait block `create project` until project exists
 
 # v0.38.0
 - Update linux build to musl